+++ src/app.ts
@@ -171,6 +171,25 @@ app.get('/set', async (req: Request, res: Response) => {
     }
 });
 
+app.get('/ext', async (req: Request, res: Response) => {
+    try {
+        const uriParameter = req.query[aniyaeHash] as string;
+        const decodedUri = Buffer.from(uriParameter, 'base64').toString('utf-8');
+
+        const response = {
+            'Esta uri ya no sera soportadad en Aniyae, hemos enviado un reporte para veridicarla': decodedUri
+        };
+        Sentry.captureException('Filelion uri encontrada, lista para clonar manualmente ' + decodedUri);
+        res.send(response);
+    } catch (error) {
+        console.error('Error generating prod-ext content :', error);
+        const response = {
+            error: 'Error generating prod-ext content '
+        };
+        res.status(500).json(response);
+    }
+});
+
 app.get('/prod-down', async (req: Request, res: Response) => {
     try {
         const uriParameter = req.query[aniyaeHash] as string;
