+++ src/index.ts
@@ -4,5 +4,5 @@ export * from './providers/prod-general';
 export * from './providers/prod-analyzer-lbry';
 export * from './providers/prod-raidenplayer';
 export * from './providers/prod-analizer-ok';
-export * from './providers/set';
+export * from './providers/prod-set';
 export * from './providers/prod-down';
